Reorganize project and use brackets in imports
